# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * HireShield API — open-source recruitment trust intelligence.
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Error {
  error: string;
}

export type FraudRisk = typeof FraudRisk[keyof typeof FraudRisk];


export const FraudRisk = {
  low: 'low',
  medium: 'medium',
  high: 'high',
  critical: 'critical',
} as const;

export type SignalSeverity = typeof SignalSeverity[keyof typeof SignalSeverity];


export const SignalSeverity = {
  info: 'info',
  low: 'low',
  medium: 'medium',
  high: 'high',
} as const;

export type SignalCategory = typeof SignalCategory[keyof typeof SignalCategory];


export const SignalCategory = {
  nlp: 'nlp',
  metadata: 'metadata',
  urgency: 'urgency',
  stylometry: 'stylometry',
  duplicate: 'duplicate',
  domain: 'domain',
  llm: 'llm',
} as const;

export interface TrustSignal {
  label: string;
  severity: SignalSeverity;
  category: SignalCategory;
  detail: string;
  /**
     * Optional numeric weight contributed to the overall score
     * @nullable
     */
  score?: number | null;
}

export interface AnalysisInput {
  /** @maxLength 300 */
  jobTitle?: string;
  /** @maxLength 200 */
  company?: string;
  /** @maxLength 320 */
  recruiterEmail?: string;
  /**
     * @maxLength 2048
     * @pattern ^https?://
     */
  jobUrl?: string;
  /**
     * @minLength 20
     * @maxLength 20000
     */
  jobDescription: string;
}

export interface AnalysisSummary {
  id: string;
  jobTitle: string;
  company: string;
  /**
     * @minimum 0
     * @maximum 100
     */
  trustScore: number;
  fraudRisk: FraudRisk;
  /**
     * @minimum 0
     * @maximum 100
     */
  ghostJobProbability: number;
  /**
     * @minimum 0
     * @maximum 100
     */
  confidenceLevel: number;
  /** @nullable */
  topSignal?: string | null;
  createdAt: string;
}

export interface Analysis {
  id: string;
  jobTitle: string;
  company: string;
  /** @nullable */
  recruiterEmail?: string | null;
  /** @nullable */
  jobUrl?: string | null;
  jobDescription: string;
  /**
     * @minimum 0
     * @maximum 100
     */
  trustScore: number;
  fraudRisk: FraudRisk;
  /**
     * @minimum 0
     * @maximum 100
     */
  ghostJobProbability: number;
  /**
     * @minimum 0
     * @maximum 100
     */
  confidenceLevel: number;
  signals: TrustSignal[];
  aiExplanation: string;
  candidateSummary: string;
  recommendedActions?: string[];
  createdAt: string;
}

export interface AnalysisExample {
  id: string;
  /** Short tag e.g. 'Classic crypto scam' */
  label: string;
  jobTitle: string;
  company: string;
  jobDescription: string;
  expectedRisk: FraudRisk;
  /** Why this example matters */
  description: string;
}

export interface RiskBreakdown {
  risk: FraudRisk;
  count: number;
}

export interface SignalCount {
  label: string;
  count: number;
}

export interface CommunityStats {
  totalAnalyses: number;
  /** Analyses with risk >= medium */
  flaggedCount: number;
  /**
     * @minimum 0
     * @maximum 100
     */
  averageTrustScore: number;
  /**
     * Percent of analyses with ghost probability >= 60
     * @minimum 0
     * @maximum 100
     */
  ghostJobShare: number;
  riskBreakdown: RiskBreakdown[];
  topSignals: SignalCount[];
  /** GitHub-style contributor count placeholder */
  contributors: number;
  lastUpdated: string;
}

export type WaitlistRole = typeof WaitlistRole[keyof typeof WaitlistRole];


export const WaitlistRole = {
  candidate: 'candidate',
  recruiter: 'recruiter',
  contributor: 'contributor',
  researcher: 'researcher',
} as const;

export interface WaitlistInput {
  /** @maxLength 320 */
  email: string;
  role?: WaitlistRole;
  /** @maxLength 200 */
  github?: string;
  /** @maxLength 1000 */
  note?: string;
}

export interface WaitlistEntry {
  id: string;
  email: string;
  role?: WaitlistRole | null;
  createdAt: string;
}

export type ListAnalysesParams = {
/**
 * @minimum 1
 * @maximum 50
 */
limit?: number;
};