# lib/api-zod/src/generated/types/analysis.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * HireShield API — open-source recruitment trust intelligence.
 * OpenAPI spec version: 0.1.0
 */
import type { FraudRisk } from './fraudRisk';
import type { TrustSignal } from './trustSignal';

export interface Analysis {
  id: string;
  jobTitle: string;
  company: string;
  /** @nullable */
  recruiterEmail?: string | null;
  /** @nullable */
  jobUrl?: string | null;
  jobDescription: string;
  /**
     * @minimum 0
     * @maximum 100
     */
  trustScore: number;
  fraudRisk: FraudRisk;
  /**
     * @minimum 0
     * @maximum 100
     */
  ghostJobProbability: number;
  /**
     * @minimum 0
     * @maximum 100
     */
  confidenceLevel: number;
  signals: TrustSignal[];
  aiExplanation: string;
  candidateSummary: string;
  recommendedActions?: string[];
  createdAt: Date;
}