# lib/api-zod/src/generated/types/analysisSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * HireShield API — open-source recruitment trust intelligence.
 * OpenAPI spec version: 0.1.0
 */
import type { FraudRisk } from './fraudRisk';

export interface AnalysisSummary {
  id: string;
  jobTitle: string;
  company: string;
  /**
     * @minimum 0
     * @maximum 100
     */
  trustScore: number;
  fraudRisk: FraudRisk;
  /**
     * @minimum 0
     * @maximum 100
     */
  ghostJobProbability: number;
  /**
     * @minimum 0
     * @maximum 100
     */
  confidenceLevel: number;
  /** @nullable */
  topSignal?: string | null;
  createdAt: Date;
}