# lib/api-zod/src/generated/types/signalCategory.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * HireShield API — open-source recruitment trust intelligence.
 * OpenAPI spec version: 0.1.0
 */

export type SignalCategory = typeof SignalCategory[keyof typeof SignalCategory];


export const SignalCategory = {
  nlp: 'nlp',
  metadata: 'metadata',
  urgency: 'urgency',
  stylometry: 'stylometry',
  duplicate: 'duplicate',
  domain: 'domain',
  llm: 'llm',
} as const;