# lib/api-zod/src/generated/types/signalSeverity.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * HireShield API — open-source recruitment trust intelligence.
 * OpenAPI spec version: 0.1.0
 */

export type SignalSeverity = typeof SignalSeverity[keyof typeof SignalSeverity];


export const SignalSeverity = {
  info: 'info',
  low: 'low',
  medium: 'medium',
  high: 'high',
} as const;