# lib/api-zod/src/generated/types/waitlistEntry.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * HireShield API — open-source recruitment trust intelligence.
 * OpenAPI spec version: 0.1.0
 */
import type { WaitlistRole } from './waitlistRole';

export interface WaitlistEntry {
  id: string;
  email: string;
  role?: WaitlistRole | null;
  createdAt: Date;
}