# lib/api-zod/src/generated/types/waitlistInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * HireShield API — open-source recruitment trust intelligence.
 * OpenAPI spec version: 0.1.0
 */
import type { WaitlistRole } from './waitlistRole';

export interface WaitlistInput {
  /** @maxLength 320 */
  email: string;
  role?: WaitlistRole;
  /** @maxLength 200 */
  github?: string;
  /** @maxLength 1000 */
  note?: string;
}