# lib/api-zod/src/generated/types/waitlistRole.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * HireShield API — open-source recruitment trust intelligence.
 * OpenAPI spec version: 0.1.0
 */

export type WaitlistRole = typeof WaitlistRole[keyof typeof WaitlistRole];


export const WaitlistRole = {
  candidate: 'candidate',
  recruiter: 'recruiter',
  contributor: 'contributor',
  researcher: 'researcher',
} as const;